requirements.txt
django
djangorestframework
markdown
requests
discord
python-dateutil
django-environ